==> 1. Importing libraries and Python data types..ipynb <==
## 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 02. Data types

x=50

x*2

y=50

x+y

(x-25)*4

x=7.82

y=2.111

x/y

a='Hello,'

b='world'

a+b

c='Strings represent textual data.'
d='They need to be enclosed in quotation marks.'

print(c+d)

print(a*100)
