==> 2. IC Data import and descriptive analysis..ipynb <==
## 0.1 Importing libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

## 0.2 Importing orders dataframe

#Importing data
df = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# setting the path
path=r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'

## 0.3 Omitting the “eval_set” column

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head(5)

## 0.4 Importing products dataframe

df_prods=pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()

The maximum value of the “aisle_id” column is 134.000000.

df_prods.dtypes

The data type of the “department_id” column is integer(int64).

df
