==> 3. Data Wrangling. Subsetting.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

#Importing data_orders
df_ords = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

#Importing data_products
df_prods= pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

## Data Wrangling

# Dropping eval_set column from df_ords
df_ords.drop(columns = ['eval_set'])

# Looking for missing values
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming Column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Changing a Variable’s Data Type
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Importing data set departments
df_dep = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\departments.csv', index_col = False)

df_dep.head()

# Transposing df_dep
df_dep.T

# Creating a new dataframe from the transposed version
df_dep_t = df_dep.T

df_dep_t

# Adding an index
df_dep_t.reset_index()

# Creating a new header
new_header = df_dep_t.iloc[0]

# Creating a new dataframe that only copies over rows beyond the first row
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Setting the header row
df_dep_t_new.columns = new_header

df_dep_t_new

## Data Dictionaries

# Creating a data_dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19')) # to show the value from the dictionary

## Subsetting

# Creating a subset
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

## Performing Task 4.4

df_ords.head()

# 2. Changing a variable’s data_type
df_ords['user_id'] = df_ords['user_id'].astype('str')

# 3. Renaming a column
df_ords.rename(columns = {'order_hour_of_day' : 'order_hour'}, inplace = True)

# 4. Distribution of orders
df_ords['order_hour'].value_counts(dropna = False)

The busiest hour for placing orders is 10.

# 5. Determining the meaning behind a value of 4 using a data dictionary.
print(data_dict.get('4'))

# 6. Creating a subset
df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast.head()

# 7. Creating subsets for dinner parties
df_alcohol = df_prods[df_prods['department_id']==5]
df_deli = df_prods[df_prods['department_id']==20]
df_beverages = df_prods[df_prods['department_id']==7]
df_meat_seafood = df_prods[df_prods['department_id']==12]

df_alcohol.head()

# Concatenating the subsets into a single DataFrame
df_dinner_parties = pd.concat([df_alcohol, df_deli, df_beverages, df_meat_seafood])

# Resetting the index is a good practice after concatenation
df_dinner_parties.reset_index(drop=True, inplace=True)

print(df_dinner_parties)

# 8. The last dataframe(df_dinner_parties) has 7650 rows x 5 columns.

# 9. Extracting all the information about the user
df_user_info = df_ords[df_ords['user_id'] == "1"] # 1 is a string because I changed it earlier.

df_user_info

# 10. Providing some details about this user’s behavior
df_user_info.describe()

# 11-12. Exporting the dataframes
path=r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
