==> 4. Data Consistency Checks.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

#Importing data_products
df_prods= pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

## Mixed-Type Data

#Creating a test df
df_test = pd.DataFrame()

#Creating a mix-type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Check for mix types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

#Changing data type
df_test['mix'] = df_test['mix'].astype('str')

## Missing Values

#Finding Missing Values
df_prods.isnull().sum()

# Creating a subset with missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

#Addressing Missing Values
df_prods.shape

#Creating a clean df
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

## Duplicates

#Looking for duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

#Creating a df with no duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Exporting the df_prods
path=r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

## Task 4.5

# Importing prepared data
df_ords = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_wrangled.csv', index_col = False)

df_ords.head()

# Run the df.describe() function on your df_ords.
df_ords.describe()

Investigating our frequency table, in particular the max and min section, it appears that days_since_prior_order has strange results. We can see that the min- 0 and max- 3, however, the 50% is 7, which is higher than the max.

# 3. Checking for mixed-type data

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
   print (col)

df_ords.dtypes

There are no mixed-type data in df_ords.

# 5.Checking for Missing Values
df_ords.isnull().sum()

# Creating a subset with missing values
df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_nan

We have 206209 rows of missing values of the days_since_prior_order column. This is not a significant number considering all records of this column. Also, we can see that the missing values have the same order_number - 1. So it can be that there is no record because the customers were making their first orders. In my opinion, we can leave it empty or impute the mode value.

# 6. Address the missing values using an appropriate method.
df_ords['order_number'].value_counts(dropna = False) 

As we see, the number of first orders is the same as the missing records. So I consider leaving this column with missing values.

# 7. Checking for duplicate values in df_ords
df_dups = df_ords[df_ords.duplicated()]

df_dups

No duplicates found in df_ords data frame.

# Exporting df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))

df_ords.describe()

df_ords.describe().astype(int)
