==> 5. Combining & Exporting Data.ipynb <==
### Combining & Exporting Data. Task 4.6

# Import libraries
import pandas as pd
import numpy as np
import os

#### 1. Importing the orders_products_combined dataframe from the pickle file

# Impotring df_ords_prods_combined
df_ords_prods_combined = pd.read_pickle(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')

#### 2. Check the shape of the imported dataframe

df_ords_prods_combined.shape

#### 3. Importing the second dataframe

# Impotring df_prods
df_prods= pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\products_checked.csv', index_col = False)

df_prods.shape

df_prods.head()

df_prods_copy = df_prods.drop(columns = ['Unnamed: 0'])

df_prods_copy

df_prods = df_prods_copy

df_prods.head()

df_ords_prods_combined.head()

df_ords_prods_combined_copy = df_ords_prods_combined.drop(columns = ['_merge'])

df_ords_prods_combined = df_ords_prods_combined_copy

#### 4. Merging the dataframes

# Merging the df_prods and df_ords_prods_combined with a merge() function, using the default option of inner join 
df_ords_prods_merge = df_prods.merge(df_ords_prods_combined, on = 'product_id', indicator = True)

# Checking the output
df_ords_prods_merge.head()

# Looking at the merge flag frequency
df_ords_prods_merge['_merge'].value_counts()

#### 5. Exporting the file

# Exporting the df_ords_prods_merge

path = r"C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis"
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
