==> 6. Combining & Exporting dataframe_merged_large.ipynb <==
## Combining & Exporting Data

# Import libraries
import pandas as pd
import numpy as np
import os

# Importing a df_ords_prior
df_ords_prior = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\orders_products_prior.csv', index_col = False)

# Check the output
df_ords_prior.head()

df_ords_prior.shape

df_ords = pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_checked.csv', index_col = False)

df_ords

df_ords_copy = df_ords.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1'])

df_ords_copy

df_ords = df_ords_copy

df_ords.shape

# Merging the two dataftrames
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# Export data to pkl

path=r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
