==> 7.  Deriving New Variables.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

#Importing data
ords_prods_merge = pd.read_pickle(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_merge.pkl')

## If-Statements with User-Defined Functions

ords_prods_merge.head()

# To avoid any potential memory issues, creating a subset
df = ords_prods_merge[:1000000]

# Creating an if-function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Designating the location for the labels
df['price_range'] = df.apply(price_label, axis=1)

# Checking the values in the new column
df['price_range'].value_counts(dropna=False)

# Checking the most expensive price
df['prices'].max()

## If-Statements with the loc() Function

# Creating conditions

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#Checking the values in the new column
df['price_range_loc'].value_counts(dropna=False)

# Applying the function for the entire df(ords_prods_merge)
ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#Checking the values in the new column
ords_prods_merge['price_range_loc'].value_counts(dropna=False)

## If-Statements with For-Loops

# the frequency of the “orders_day_of_week” column
ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

# Creating a for-loop for orders_day_of_week column
result = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# Creating a new column
ords_prods_merge['busiest_day']=result

ords_prods_merge['busiest_day'].value_counts(dropna=False)

## Task

### 1. Update the “Busiest day” to “Busiest days”.

# Creating a new for-loop to identify two busiest days (0, 1), slowest days (4, 3), or regularly busy (all other days).
result_2 = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")
  elif value == 4 or value == 3:
    result_2.append("Slowest days")
  else:
    result_2.append("Regularly busy")

result_2

# Creating a new column
ords_prods_merge['busiest_days']=result_2

ords_prods_merge['busiest_days'].value_counts(dropna=False)

# Check output of ords_prods_merge with new "busiest days" column.

ords_prods_merge.head()

The total value number for the ords_prods_merge dataframe is equal to the total sum of all groupings ("Regularly busy" + "Busiest days" + "Slowest days").

### 2. Create a new column containing labels called “busiest_period_of_day.”

# Checking value counts in "order_hour_of_day" column.

ords_prods_merge['order_hour'].value_counts(dropna=False)

“Most orders” = 10, 11, 14, 15, 13, 12, 16, and 9;
“Average orders”= 17, 8, 18, 19, 20, 7, 21, 22;
“Fewest orders”= 23, 6, 0, 1, 5, 2, 4, and 3.

# Creating a for-loop to identify two busiest hours of the day.
result_3 = []

for value in ords_prods_merge["order_hour"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9]:
    result_3.append("Most orders")
  elif value in [23, 6, 0, 1, 5, 2, 4, 3]:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

result_3

# Creating a new column
ords_prods_merge['busiest_period_of_day']=result_3

ords_prods_merge['busiest_period_of_day'].value_counts(dropna=False)

# Check the output 
ords_prods_merge.head()

# Exporting df
path = r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'
ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))
