==> 8. Grouping Data & Aggregating Variables P1.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

#Importing data
ords_prods_merge = pd.read_pickle(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged_derived.pkl')

#Creating a subset
df = ords_prods_merge[:1000000]

df.shape

df.head()

## Grouping Data with pandas

#Split the data into groups based on some criteria
df.groupby('product_name')

## Aggregating Data with agg()

### Performing a Single Aggregation

df.groupby('department_id').agg({'order_number': ['mean']})

#Replacing with a command that uses the mean() function to achieve the same results
df.groupby('department_id')['order_number'].mean()

### Performing Multiple Aggregations

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

## Aggregating Data with transform()

# Creating a new column “max_order”
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(15)

## Deriving Columns with loc()

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Checking the output
ords_prods_merge['loyalty_flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)
