==> 9. Grouping Data & Aggregating Variables. P2.ipynb <==
## 1. Preparing the notebook

# Import libraries
import pandas as pd
import numpy as np
import os

#Importing data
ords_prods_merge = pd.read_pickle(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged_derived.pkl')

## 2. Aggregating Data with agg()

# Grouping and aggregating the df 
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

## 3. Analyzing the result.

As we can see, the maximum and minimum values are the same as in the subset because they were present in the subset, but the average values are different because the dataframe of a million rows did not include all records.

## 4. Aggregating Data with transform() and loc()

# Creating a new column “max_order”
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(15)

# Deriving Columns with loc()
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Checking the output
ords_prods_merge['loyalty_flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

## 5. Determining is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean']})

ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['sum']})

## 6. Create a spending flag for each user based on the average price across all their orders using the criteria.

# Creating a new column “avg_price”
ords_prods_merge['avg_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.head(15)

# Deriving the column with loc()
ords_prods_merge.loc[ords_prods_merge['avg_price'] < 10, 'spending flag'] = 'Low spender'

ords_prods_merge.loc[ords_prods_merge['avg_price'] >= 10, 'spending flag'] = 'High spender'

# Checking the output
ords_prods_merge['spending flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'spending flag', 'avg_price']].head(10)

## 7. Creating an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column.

# Creating a new column “median_order_frequency”
ords_prods_merge['median_order_frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.head(15)

# Deriving the column with loc()
ords_prods_merge.loc[ords_prods_merge['median_order_frequency'] > 20, 'frequency flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['median_order_frequency'] > 10) & (ords_prods_merge['median_order_frequency'] <= 20), 'frequency flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['median_order_frequency'] <= 10, 'frequency flag'] = 'Frequent customer'

# Checking the output
ords_prods_merge['frequency flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'frequency flag', 'median_order_frequency']].head(15)

#Exporting the df
path = r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'
ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_aggregated.pkl'))
