==> 10. Customer data Check.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

#Importing customers data
df_cust= pd.read_csv(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Original Data\customers.csv', index_col = False)

df_cust.head(20)

df_cust.shape

df_cust.describe()

## 1. Wrangling the dataframe

### 1.1 Renaming 

# Renaming Columns
df_cust.rename(columns = {'First Name' : 'first_name','Last Name' : 'last_name', 'State' : 'state','Gender' : 'gender', 'Age' : 'age'}, inplace = True)

df_cust.head()

### 1.2 Changing data type

#Changing the data type of user_id
df_cust['user_id'] = df_cust['user_id'].astype('str')

df_cust.info() #before it was int

## 2. The fundamental data quality and consistency check

### 2.1 Checking for the Mixed-Type Data

# Check for mix types
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

#Changing data type
df_cust['first_name'] = df_cust['first_name'].astype('str')

### 2.2 Missing Values

#Looking for Missing Values
df_cust.isnull().sum()

### 2.3 Duplicates

#Looking for duplicates
df_dups = df_cust[df_cust.duplicated()]

df_dups

#### RESULTS: 
#### There are no Missing Values and Duplicates in the df_cust. The Mixed-Type Data was identified and changed to the string data type.

## 3. Combine customer data with the rest of prepared Instacart data.

### 3.1 Import orders-product pickle file

# Importing the data
df_ords_prods = pd.read_pickle(r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_aggregated.pkl')

df_ords_prods.head()

df_ords_prods.shape

df_cust.head()

df_cust.shape

### 3.2 Changing data type of the key column

df_ords_prods.info()

#Changing data type of the key column
df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')

### 3.3 Preparing the dataframe

# Droping the _merge column in df_ords_prods
df_ords_prods.drop(columns=['_merge'], inplace=True)

### 3.4 Merging the dataframes

# Merging the df_ords_prods and df_cust
df_ords_prods_cust = df_ords_prods.merge(df_cust, on = 'user_id', indicator = True)

# Checking the output
df_ords_prods_cust['_merge'].value_counts()

df_ords_prods_cust.head()

### 3.5 Exporting the data

# Export data to pkl

path=r'C:\Users\ahamr\Desktop\12-2023 Instacart Basket Analysis'

df_ords_prods_cust.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_customers.pkl'))
